'sweetalert2'

+++ src/views/dashboardUser.js
@@ -1,7 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { Table, Button } from 'react-bootstrap';
 import styles from '../styles/user.module.css';
-import HiOutlinePencilSquare from 'react-icons/hi2'
+import Swal from 'sweetalert2';
+import withReactContent from 'sweetalert2-react-content'
 import { useNavigate, Link } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import {
@@ -18,20 +19,51 @@ const DashboardUser = () => {
     const [currentIndex, setCurrentIndex] = useState(-1);
 
     const dispatch = useDispatch();
+    const mySwal = withReactContent(Swal);
 
 
     const { getUserListResult, getUserLoading, getUserError, userDeleteResult } =
         useSelector((state) => state.userReducer);
 
     useEffect(() => {
+
         dispatch(getUserList());
-    }, [])
+
+
+    }, [dispatch])
 
 
     const deleteById = (id, e) => {
         e.preventDefault();
         console.log(e)
-        dispatch(deleteUser(id))
+        mySwal.fire({
+            title: "Konfirmasi?",
+            text: "Anda yakin akan menghapusnya",
+            icon: "warning",
+            dangerMode: true,
+            confirmButtonColor: '#3085d6',
+            showCancelButton: true,
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Hapus!',
+            cancelButtonText: 'Batal'
+
+        })
+            .then((result) => {
+                if (result.isConfirmed) {
+                    dispatch(deleteUser(id))
+                    mySwal.fire({
+                        position: 'center',
+                        title: "Yeay",
+                        text: "Berhasil menghapus",
+                        icon: "success",
+                        timer: 2000,
+                        showConfirmButton: false,
+                    }).then(() => {
+                        window.location.reload()
+                    })
+
+                }
+            })
 
     }
     let navigate = useNavigate();
@@ -47,7 +79,7 @@ const DashboardUser = () => {
             </div>
             <div className={styles.headerBody2}>
                 <Button variant="primary" size="sm" active className={styles.btnExport} onClick={tambahUser}>
-                    Tambah
+                    +
                 </Button>
                 {/* <Button variant="primary" size="sm" active className={styles.btnExport}>
                     Export as PDF
@@ -56,14 +88,14 @@ const DashboardUser = () => {
                     Export as Excel
                 </Button> */}
             </div>
-            <Table striped bordered hover>
-                <thead className={styles.tableBody}>
+            <Table responsive className={styles.bodyTable}>
+                <thead className={styles.theadBody}>
                     <tr>
                         <th>No</th>
                         <th>Nama Lengkap</th>
                         <th>Username</th>
                         <th>Email</th>
-                        <th>Password</th>
+
                         <th>Nomor Whatsapp</th>
                         <th>Alamat</th>
                         <th>Aksi</th>
@@ -77,12 +109,11 @@ const DashboardUser = () => {
                                 <td>{e.fullName}</td>
                                 <td>{e.username ? (e.username) : "-"}</td>
                                 <td>{e.email}</td>
-                                <td>{e.password}</td>
                                 <td>{e.noWa ? (e.noWa) : "-"}</td>
                                 <td>{e.alamat ? (e.alamat) : "-"}</td>
                                 <td className={styles.tdBtn}>
-                                    <Button variant="success" size="sm"><Link to={"/detail-user/" + e._id}>Edit</Link></Button>
-                                    <Button variant="danger" size="sm" onClick={(x) => deleteById(e._id, x)}>Hapus</Button>
+                                    <Button variant="success" size="sm"><Link to={"/detail-user/" + e._id} className={styles.linkTo}>Edit</Link></Button>
+                                    <Button variant="danger" size="sm" onClick={(x) => deleteById(e._id, x)} className={styles.linkTo}>Hapus</Button>
                                 </td>
                             </tr>
                         )
