+++ src/views/wisata.js
@@ -1,14 +1,17 @@
 import React, { useEffect } from 'react'
 import { Table, Button } from 'react-bootstrap';
-import styles from '../styles/user.module.css';
+import styles from '../styles/wisata.module.css';
 import HiOutlinePencilSquare from 'react-icons/hi2'
 import { useNavigate, Link } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import { getWisataList, deleteWisata } from "../actions/wisataAction";
+import Swal from 'sweetalert2';
+import withReactContent from 'sweetalert2-react-content';
 
 const Wisata = () => {
 
     const dispatch = useDispatch();
+    const mySwal = withReactContent(Swal);
 
     const { getWisataListResult, getWisataLoading, getWisataError } =
         useSelector((state) => state.wisataReducer);
@@ -26,7 +29,34 @@ const Wisata = () => {
     const deleteById = (id, e) => {
         e.preventDefault();
         console.log(e)
-        dispatch(deleteWisata(id))
+        mySwal.fire({
+            title: "Konfirmasi?",
+            text: "Anda yakin akan menghapusnya",
+            icon: "warning",
+            dangerMode: true,
+            confirmButtonColor: '#3085d6',
+            showCancelButton: true,
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Hapus!',
+            cancelButtonText: 'Batal'
+
+        })
+            .then((result) => {
+                if (result.isConfirmed) {
+                    dispatch(deleteWisata(id))
+                    mySwal.fire({
+                        position: 'center',
+                        title: "Yeay",
+                        text: "Berhasil menghapus",
+                        icon: "success",
+                        timer: 2000,
+                        showConfirmButton: false,
+                    }).then(() => {
+                        window.location.reload()
+                    })
+
+                }
+            })
 
     }
 
@@ -37,17 +67,17 @@ const Wisata = () => {
             </div>
             <div className={styles.headerBody2}>
                 <Button variant="primary" size="sm" active className={styles.btnExport} onClick={tambahWisata}>
-                    Tambah
+                    +
                 </Button>
-                <Button variant="primary" size="sm" active className={styles.btnExport}>
+                {/* <Button variant="primary" size="sm" active className={styles.btnExport}>
                     Export as PDF
                 </Button>
                 <Button variant="primary" size="sm" active className={styles.btnExport}>
                     Export as Excel
-                </Button>
+                </Button> */}
             </div>
-            <Table striped bordered hover>
-                <thead className={styles.tableBody}>
+            <Table responsive className={styles.bodyTable}>
+                <thead className={styles.theadBody}>
                     <tr>
                         <th>No</th>
                         <th>Nama Wisata</th>
@@ -68,8 +98,8 @@ const Wisata = () => {
                                 <td>{e.rating ? (e.rating) : "-"}</td>
 
                                 <td className={styles.tdBtn}>
-                                    <Button variant="success" size="sm"><Link to={"/detail-wisata/" + e._id}>Edit</Link></Button>
-                                    <Button variant="danger" size="sm" onClick={(x) => deleteById(e._id, x)}>Hapus</Button>
+                                    <Button variant="success" size="sm"><Link to={"/detail-wisata/" + e._id} className={styles.linkTo}>Edit</Link></Button>
+                                    <Button variant="danger" size="sm" onClick={(x) => deleteById(e._id, x)} className={styles.linkTo}>Hapus</Button>
                                 </td>
                             </tr>
                         )
